First round of merge fixes
