+++ src/lqc/config/file_config.py
@@ -26,8 +26,7 @@ class FileConfig:
         formatted_timestamp = timestamp.strftime(timestamp_format)
         custom_dir = os.path.join(self.bug_report_file_dir, custom_folder)
         if not os.path.exists(custom_dir):
-            os.makedirs(custom_dir)
-
+            os.makedirs(custom_dir)        
         dirpath = os.path.join(custom_dir, f"{custom_folder}-{formatted_timestamp}")
         if not os.path.exists(dirpath):
             os.makedirs(dirpath)
