+++ src/lqc_selenium/report/bug_report_helper.py
@@ -24,25 +24,28 @@ def save_bug_report(
     sorting_seconds=None,
     true_minification_seconds=None,
 ):
+    # If there is no minified subject due rule matching then set it to the prerun subject.
     if minified_run_subject is None:
         minified_run_subject = prerun_subject
 
+    # no group path, create a new bug folder in the default location
     if path is None:
-        bug_folder = FileConfig().getCustomTimestampPath()
-        group_path = None
+        raise ValueError("Path must be provided to save bug report")
     else:
+        #create a new bug folder inside the provided path if the path is a bug group, otherwise use the provided path as the bug folder
         group_path = (
             path
             if os.path.basename(os.path.normpath(path)).startswith("bug-group-")
             else None
         )
+
         if group_path is not None:
             bug_folder = os.path.join(
                 path,
                 f"bug-{datetime.now().strftime('%Y%m%d%H%M%S%f')}",
             )
         else:
-            bug_folder = path
+            raise ValueError(f"Provided path {path} is not a valid bug group path")
 
     matched_rule_folder = group_path
 
@@ -50,14 +53,17 @@ def save_bug_report(
 
     os.makedirs(bug_folder, exist_ok=True)
 
+    # Copy the original file
     bug_filepath = os.path.join(bug_folder, "original_bug.html")
     shutil.copy(original_filepath, bug_filepath)
 
+    # Copy the minimized bug
     minified_bug = os.path.join(bug_folder, "minified_bug.html")
     print(f"Saving minimized bug to {minified_bug}")
     save_as_web_page(minified_run_subject, minified_bug, run_result=run_result)
     copyExternalJSFiles(bug_folder)
 
+    # Custom bug helper file - JSON file
     styles_used = list(minified_run_subject.all_style_names())
     styles_used.sort()
     styles_used_string = ",".join(styles_used)
@@ -90,17 +96,18 @@ def save_bug_report(
         "sorting_seconds": sorting_seconds,
         "true_minification_seconds": true_minification_seconds,
     }
-
+    
     if isinstance(run_result, RunResultLayoutBug):
         json_data["differences"] = run_result.element_dimensions
 
     json_data_filepath = os.path.join(bug_folder, "data.json")
-    with open(json_data_filepath, "w", encoding="utf-8") as f:
+    with open(json_data_filepath, "w") as f:
         f.write(json.dumps(json_data, indent=4, default=lambda o: o.__dict__))
 
     if group_path is not None:
         rule = recompute_bug_group_artifacts(group_path)
         if rule is None:
+            raise ValueError(f"Could not find rule for bug group at {group_path}")
             moved_paths = dissolve_bug_group(group_path)
             moved_bug_folder = moved_paths.get(os.path.abspath(bug_folder))
             if moved_bug_folder is not None:
